update Comment Service to add playground path in all endpoints

+++ spring/src/main/java/com/playground/controllers/CommentController.java
@@ -49,19 +49,25 @@ public class CommentController {
     /**
      * [GET] Return one comment by id
      *
-     * @param id int
+     * @param playgroundId int
+     * @param commentId int
      *
      * @return ResponseEntity
      *
+     * @throws ResourceNotFoundException Playground not found
      * @throws ResourceNotFoundException Comment not found
      */
-    @GetMapping(value = "/comments/{id}", produces = "application/json")
-    public ResponseEntity<?> getComment(@PathVariable("id") int id) throws ResourceNotFoundException {
-        Comment comment = commentService.getComment(id);
+    @GetMapping(value = "/playgrounds/{playgroundId}/comments/{commentId}", produces = "application/json")
+    public ResponseEntity<?> getComment(@PathVariable("playgroundId") int playgroundId, @PathVariable("commentId") int commentId) throws ResourceNotFoundException {
 
-        if (comment == null) {
-            throw new ResourceNotFoundException("Comment with id " + id + " not found");
+        Playground playground = playgroundService.getPlayground(playgroundId);
+
+        if (playground == null) {
+            throw new ResourceNotFoundException("Playground with id " + playgroundId + " not found");
         }
+
+        Comment comment = commentService.getCommentByPlayground(playground, commentId);
+        if (comment == null) throw new ResourceNotFoundException("Comment with id " + commentId + " not found for this playground");
 
         return new ResponseEntity<>(comment, HttpStatus.OK);
     }
@@ -69,73 +75,104 @@ public class CommentController {
     /**
      * [POST] Create a comment and return it
      *
+     * @param playgroundId int
      * @param comment Comment
      *
      * @return ResponseEntity
      */
-    @PostMapping(value = "/comments", consumes = "application/json")
-    public ResponseEntity<Comment> createComment(@RequestBody Comment comment) {
-        return new ResponseEntity<>(commentService.createComment(comment), HttpStatus.CREATED);
+    @PostMapping(value = "/playgrounds/{playgroundId}/comments", consumes = "application/json")
+    public ResponseEntity<Comment> createComment(@PathVariable("playgroundId") int playgroundId, @RequestBody Comment comment) {
+        Playground playground = playgroundService.getPlayground(playgroundId);
+
+        if (playground == null) {
+            throw new ResourceNotFoundException("Playground with id " + playgroundId + " not found");
+        }
+
+        return new ResponseEntity<>(commentService.createComment(playground, comment), HttpStatus.CREATED);
     }
 
     /**
      * [PUT] Update a comment and return it
      *
-     * @param id int
+     * @param playgroundId int
+     * @param commentId int
      * @param comment Comment
      *
      * @return ResponseEntity
      *
+     * @throws ResourceNotFoundException Playground not found
      * @throws ResourceNotFoundException Comment not found
      */
-    @PutMapping("/comments/{id}")
-    public ResponseEntity<Comment> updateComment(@PathVariable("id") int id, @RequestBody Comment comment) throws ResourceNotFoundException {
-        Comment currentComment = commentService.getComment(id);
+    @PutMapping("/playgrounds/{playgroundId}/comments/{commentId}")
+    public ResponseEntity<Comment> updateComment(@PathVariable("playgroundId") int playgroundId, @PathVariable("commentId") int commentId, @RequestBody Comment comment) throws ResourceNotFoundException {
 
-        if (currentComment == null) {
-            throw new ResourceNotFoundException("Comment with id " + id + " not found");
+        Playground playground = playgroundService.getPlayground(playgroundId);
+
+        if (playground == null) {
+            throw new ResourceNotFoundException("Playground with id " + playgroundId + " not found");
         }
 
-        return new ResponseEntity<>(commentService.updateComment(id, comment), HttpStatus.OK);
+        Comment currentComment = commentService.getCommentByPlayground(playground, commentId);
+
+        if (currentComment == null) {
+            throw new ResourceNotFoundException("Comment with id " + commentId + " not found for this playground");
+        }
+
+        return new ResponseEntity<>(commentService.updateComment(commentId, comment), HttpStatus.OK);
     }
 
     /**
      * [DELETE] Delete a comment
      *
-     * @param id int
+     * @param playgroundId playgroundId
+     * @param commentId int
      *
      * @return ResponseEntity
      *
+     * @throws ResourceNotFoundException Playground not found
      * @throws ResourceNotFoundException Comment not found
      */
-    @DeleteMapping("/comments/{id}")
-    public ResponseEntity<?> deleteComment(@PathVariable("id") int id) throws ResourceNotFoundException {
-        Comment currentComment = commentService.getComment(id);
+    @DeleteMapping("/playgrounds/{playgroundId}/comments/{commentId}")
+    public ResponseEntity deleteComment(@PathVariable("playgroundId") int playgroundId, @PathVariable("commentId") int commentId) throws ResourceNotFoundException {
+        Playground playground = playgroundService.getPlayground(playgroundId);
+
+        if (playground == null) {
+            throw new ResourceNotFoundException("Playground with id " + playgroundId + " not found");
+        }
+
+        Comment currentComment = commentService.getCommentByPlayground(playground, commentId);
 
         if (currentComment == null) {
-            throw new ResourceNotFoundException("Comment with id " + id + " not found");
+            throw new ResourceNotFoundException("Comment with id " + commentId + " not found for this playground");
         }
 
         commentService.deleteComment(currentComment);
 
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        return new ResponseEntity(HttpStatus.NO_CONTENT);
     }
 
     /**
      * [PUT] Archived a comment
      *
-     * @param id int
+     * @param playgroundId playgroundId
+     * @param commentId int
      *
      * @return ResponseEntity
      *
      * @throws ResourceNotFoundException Comment not found
      */
-    @PutMapping("/comments/archived/{id}")
-    public ResponseEntity<Comment> archivedComment(@PathVariable("id") int id) throws ResourceNotFoundException {
-        Comment currentComment = commentService.getComment(id);
+    @PutMapping("/playgrounds/{playgroundId}/comments/archived/{commentId}")
+    public ResponseEntity<Comment> archivedComment(@PathVariable("playgroundId") int playgroundId, @PathVariable("commentId") int commentId) throws ResourceNotFoundException {
+        Playground playground = playgroundService.getPlayground(playgroundId);
+
+        if (playground == null) {
+            throw new ResourceNotFoundException("Playground with id " + playgroundId + " not found");
+        }
+
+        Comment currentComment = commentService.getCommentByPlayground(playground, commentId);
 
         if (currentComment == null) {
-            throw new ResourceNotFoundException("Comment with id " + id + " not found");
+            throw new ResourceNotFoundException("Comment with id " + commentId + " not found for this playground");
         }
 
         return new ResponseEntity<>(commentService.archivedComment(currentComment), HttpStatus.OK);
@@ -151,14 +188,14 @@ public class CommentController {
      * @throws ResourceNotFoundException Playground not found
      */
     @GetMapping(value = "/playgrounds/{playgroundId}/comments", produces = "application/json")
-    public ResponseEntity<List<Comment>> search(@PathVariable(value = "playgroundId") int playgroundId) throws ResourceNotFoundException {
+    public ResponseEntity<List<Comment>> getCommentsByPlaygroundId(@PathVariable(value = "playgroundId") int playgroundId) throws ResourceNotFoundException {
         Playground playground = playgroundService.getPlayground(playgroundId);
 
         if (playground == null) {
             throw new ResourceNotFoundException("Playground with id " + playgroundId + " not found");
         }
 
-        List<Comment> listComments = commentService.getCommentByPlaygroundId(playground);
+        List<Comment> listComments = commentService.getCommentsByPlayground(playground);
 
         return new ResponseEntity<>(listComments, HttpStatus.OK);
     }
